manual test case and automated test script
This is a quality assurance assessment involving the discrepancies between writing a test script and writing a test case. This two different methods are unique in their own way and this assessment can also reiterate the importance of manual test cases. this assessment shows how the two ways of testing can be used.